LearnNCS/LEC2_Python_ParseFile/requirements.txt
lief==0.12.3
python-magic-bin==0.4.14